Fix race condition when processing ping-request
When making a ping-request,
it was possible for response handler to execute before
the ping-request Id to be added m_elevatedPings when
handling responses from low-latency VPN servers.
This resulted in response being ignored and "lost".

The primary side-effect of this (now fixed) bug is that
a VPN server would be marked as invalid because LastPingResult was never set.
Because LastPingTest is already set, Eddie will not attempt to retry the ping for 180 seconds.
While the ConnectionInfo for the VPN server is invalid,
it will block users from connecting to VPN servers with the message "Waiting for latency tests" and
without reattempting the connection.